+++ CHANGELOG.md
@@ -4,9 +4,8 @@
 
 ### Added
 
-- Improved Falcon integration: `set_api()`  is no longer required, tracing
-  is automatically disabled if installation fails, and callable class-based
-  responders are supported.
+- Improved Falcon integration: tracing is automatically disabled if
+  installation fails, and callable class-based responders are supported.
 
 ## [2.10.0] 2020-01-02